Added some input data
I added three different test values for the items and invoices tables. As a side-note, 'Customer ID' examples are in a string format, while the current variable is 'int'. This will cause an error if the associated column isn't changed to a string. I did not do so to prevent a merge conflict; the associated change should be in Ernie's branch. If this branch is merged and his own is not, errors will occur.